Move Related Products and Signup to coroutine and add more features

+++ app/src/main/java/school/cactus/succulentshop/api/product/Image.kt
@@ -2,6 +2,6 @@ package school.cactus.succulentshop.api.product
 
 data class Image(
     val url: String,
-    val width: Long,
-    val height: Long
+    val width: Int,
+    val height: Int
 )
