+++ app/src/main/java/school/cactus/succulentshop/api/product/ImageInfo.kt
@@ -3,7 +3,7 @@ package school.cactus.succulentshop.api.product
 data class ImageInfo(
     val id: Long,
     val name: String,
-    val width: Long,
-    val height: Long,
+    val width: Int,
+    val height: Int,
     val formats: ImageFormats,
 )
